+++ CopilotKit/examples/next-openai/src/app/presentation/page.tsx
@@ -2,18 +2,23 @@
 
 import { useCopilotContext } from "@copilotkit/react-core";
 import { CopilotTask } from "@copilotkit/react-core";
-import {
-  CopilotKit,
-  useMakeCopilotActionable,
-  useMakeCopilotReadable,
-} from "@copilotkit/react-core";
+import { CopilotKit, useMakeCopilotReadable } from "@copilotkit/react-core";
 import { CopilotSidebar } from "@copilotkit/react-ui";
 import { useEffect, useState } from "react";
 import Markdown from "react-markdown";
 import "./styles.css";
+import { useCopilotAction } from "@copilotkit/react-core";
 
 let globalAudio: any = undefined;
 let globalAudioEnabled = false;
+
+function enableGlobalAudio() {
+  if (!globalAudioEnabled) {
+    globalAudio.play();
+    globalAudio.pause();
+  }
+  globalAudioEnabled = true;
+}
 
 const Demo = () => {
   return (
@@ -26,11 +31,7 @@ const Demo = () => {
         }}
         clickOutsideToClose={false}
         onSubmitMessage={async (message) => {
-          if (!globalAudioEnabled) {
-            globalAudio.play();
-            globalAudio.pause();
-          }
-          globalAudioEnabled = true;
+          enableGlobalAudio();
         }}
       >
         <Presentation />
@@ -53,55 +54,48 @@ const Presentation = () => {
 
   useMakeCopilotReadable("This is the current slide: " + JSON.stringify(state));
 
-  useMakeCopilotActionable(
-    {
-      name: "presentSlide",
-      description:
-        "Present a slide in the presentation you are giving. Call this function multiple times to present multiple slides.",
-      argumentAnnotations: [
-        {
-          name: "markdown",
-          type: "string",
-          description:
-            "The text to display in the presentation slide. Use simple markdown to outline your speech, like a headline, lists, paragraphs, etc.",
-          required: true,
-        },
-        {
-          name: "backgroundImage",
-          type: "string",
-          description: "What to display in the background of the slide (i.e. 'dog' or 'house').",
-          required: true,
-        },
-        {
-          name: "speech",
-          type: "string",
-          description: "An informative speech about the current slide.",
-          required: true,
-        },
-      ],
-
-      implementation: async (markdown, backgroundImage, speech) => {
-        setState({
-          markdown,
-          backgroundImage,
-        });
-
-        console.log("Presenting slide: ", markdown, backgroundImage, speech);
-
-        const encodedText = encodeURIComponent(speech);
-        const url = `/api/tts?text=${encodedText}`;
-        globalAudio.src = url;
-        await globalAudio.play();
-        await new Promise<void>((resolve) => {
-          globalAudio.onended = function () {
-            resolve();
-          };
-        });
-        await new Promise((resolve) => setTimeout(resolve, 500));
+  useCopilotAction({
+    name: "presentSlide",
+    description:
+      "Present a slide in the presentation you are giving. Call this function multiple times to present multiple slides.",
+    parameters: [
+      {
+        name: "markdown",
+        type: "string",
+        description:
+          "The text to display in the presentation slide. Use simple markdown to outline your speech, like a headline, lists, paragraphs, etc.",
       },
+      {
+        name: "backgroundImage",
+        type: "string",
+        description: "What to display in the background of the slide (i.e. 'dog' or 'house').",
+      },
+      {
+        name: "speech",
+        type: "string",
+        description: "An informative speech about the current slide.",
+      },
+    ],
+    handler: async ({ markdown, speech, backgroundImage }) => {
+      setState({
+        markdown,
+        backgroundImage,
+      });
+
+      console.log("Presenting slide: ", markdown, backgroundImage, speech);
+
+      const encodedText = encodeURIComponent(speech);
+      const url = `/api/tts?text=${encodedText}`;
+      globalAudio.src = url;
+      await globalAudio.play();
+      await new Promise<void>((resolve) => {
+        globalAudio.onended = function () {
+          resolve();
+        };
+      });
+      await new Promise((resolve) => setTimeout(resolve, 500));
     },
-    [],
-  );
+  });
 
   const randomSlideTask = new CopilotTask({
     instructions: "Make a random slide related to the current topic",
@@ -119,6 +113,7 @@ const Presentation = () => {
         className={`absolute bottom-0 left-0 mb-4 ml-4 bg-blue-500 text-white font-bold py-2 px-4 rounded
         ${randomSlideTaskRunning ? "opacity-50 cursor-not-allowed" : "hover:bg-blue-700"}`}
         onClick={async () => {
+          enableGlobalAudio();
           try {
             setRandomSlideTaskRunning(true);
             await randomSlideTask.run(context);