+++ CopilotKit/examples/next-openai/CHANGELOG.md
@@ -1,6 +1,6 @@
 # web
 
-## 1.17.0-mme-assistant-api.0
+## 1.17.0-mme-experimental-actions.1
 
 ### Minor Changes
 
@@ -9,11 +9,26 @@
 ### Patch Changes
 
 - Updated dependencies
-  - @copilotkit/react-core@0.21.0-mme-assistant-api.0
-  - @copilotkit/backend@0.5.0-mme-assistant-api.0
-  - @copilotkit/shared@0.5.0-mme-assistant-api.0
-  - @copilotkit/react-textarea@0.31.0-mme-assistant-api.0
-  - @copilotkit/react-ui@0.18.0-mme-assistant-api.0
+  - @copilotkit/react-textarea@0.31.0-mme-experimental-actions.1
+  - @copilotkit/react-core@0.21.0-mme-experimental-actions.1
+  - @copilotkit/react-ui@0.18.0-mme-experimental-actions.1
+  - @copilotkit/backend@0.5.0-mme-experimental-actions.1
+  - @copilotkit/shared@0.5.0-mme-experimental-actions.1
+
+## 1.17.0-mme-experimental-actions.0
+
+### Minor Changes
+
+- QA new action type
+
+### Patch Changes
+
+- Updated dependencies
+  - @copilotkit/react-textarea@0.31.0-mme-experimental-actions.0
+  - @copilotkit/react-core@0.21.0-mme-experimental-actions.0
+  - @copilotkit/react-ui@0.18.0-mme-experimental-actions.0
+  - @copilotkit/backend@0.5.0-mme-experimental-actions.0
+  - @copilotkit/shared@0.5.0-mme-experimental-actions.0
 
 ## 1.16.1