+++ CopilotKit/packages/eslint-config-custom/CHANGELOG.md
@@ -1,10 +1,16 @@
 # eslint-config-custom
 
-## 0.6.0-mme-assistant-api.0
+## 0.6.0-mme-experimental-actions.1
 
 ### Minor Changes
 
 - Alpha
+
+## 0.6.0-mme-experimental-actions.0
+
+### Minor Changes
+
+- QA new action type
 
 ## 0.5.1