+++ CopilotKit/packages/react-ui/CHANGELOG.md
@@ -1,6 +1,6 @@
 # ui
 
-## 0.18.0-mme-assistant-api.0
+## 0.18.0-mme-experimental-actions.1
 
 ### Minor Changes
 
@@ -9,8 +9,20 @@
 ### Patch Changes
 
 - Updated dependencies
-  - @copilotkit/react-core@0.21.0-mme-assistant-api.0
-  - @copilotkit/shared@0.5.0-mme-assistant-api.0
+  - @copilotkit/react-core@0.21.0-mme-experimental-actions.1
+  - @copilotkit/shared@0.5.0-mme-experimental-actions.1
+
+## 0.18.0-mme-experimental-actions.0
+
+### Minor Changes
+
+- QA new action type
+
+### Patch Changes
+
+- Updated dependencies
+  - @copilotkit/react-core@0.21.0-mme-experimental-actions.0
+  - @copilotkit/shared@0.5.0-mme-experimental-actions.0
 
 ## 0.17.1