+++ CopilotKit/packages/tsconfig/CHANGELOG.md
@@ -1,10 +1,16 @@
 # tsconfig
 
-## 0.10.0-mme-assistant-api.0
+## 0.10.0-mme-experimental-actions.1
 
 ### Minor Changes
 
 - Alpha
+
+## 0.10.0-mme-experimental-actions.0
+
+### Minor Changes
+
+- QA new action type
 
 ## 0.9.1